requirements.txt
pandas
numpy
scikit-learn
streamlit
requestspython-dotenv